feat: Parallel Compact

+++ file/file.go
@@ -18,7 +18,7 @@ import "io"
 
 // Options
 type Options struct {
-	FID      uint32
+	FID      uint64
 	FileName string
 	Dir      string
 	Path     string
